+++ ALVRClient/Model/ViewModel.swift
@@ -15,4 +15,7 @@ class ViewModel {
     
     // Client
     var isShowingClient: Bool = false
+
+    // Server
+    var alvrInitialized: Bool = false
 }
