+++ ALVRClient/WorldTracker.swift
@@ -1,0 +1,95 @@
+//
+//  WorldTracker.swift
+//  ALVRClient
+//
+//
+
+import Foundation
+import ARKit
+import CompositorServices
+
+class WorldTracker {
+    static let shared = WorldTracker()
+    
+    let arSession: ARKitSession!
+    let worldTracking: WorldTrackingProvider!
+    
+    var deviceAnchorsLock = NSObject()
+    var deviceAnchorsQueue = [UInt64]()
+    var deviceAnchorsDictionary = [UInt64: simd_float4x4]()
+    
+    static let maxPrediction = 30 * NSEC_PER_MSEC
+    static let deviceIdHead = alvr_path_string_to_id("/user/head")
+
+    
+    init(arSession: ARKitSession = ARKitSession(), worldTracking: WorldTrackingProvider = WorldTrackingProvider()) {
+        self.arSession = arSession
+        self.worldTracking = worldTracking
+    }
+    
+    func initializeAr() async  {
+        do {
+            try await arSession.run([worldTracking])
+        } catch {
+            fatalError("Failed to initialize ARSession")
+        }
+    }
+    
+    // TODO: figure out how stable Apple's predictions are into the future
+    
+    func sendTracking(targetTimestamp: Double) {
+        let targetTimestamp = CACurrentMediaTime() + Double(min(alvr_get_head_prediction_offset_ns(), WorldTracker.maxPrediction)) / Double(NSEC_PER_SEC)
+        var targetTimestampWalkedBack = targetTimestamp
+        var deviceAnchor:DeviceAnchor? = nil
+        
+        // Predict as far into the future as Apple will allow us.
+        for i in 0...20 {
+            deviceAnchor = worldTracking.queryDeviceAnchor(atTimestamp: targetTimestampWalkedBack)
+            if deviceAnchor != nil {
+                break
+            }
+            targetTimestampWalkedBack -= (5/1000.0)
+        }
+        
+        // Fallback.
+        if deviceAnchor == nil {
+            targetTimestampWalkedBack = CACurrentMediaTime()
+            deviceAnchor = worldTracking.queryDeviceAnchor(atTimestamp: targetTimestamp)
+        }
+
+        // Well, I'm out of ideas.
+        guard let deviceAnchor = deviceAnchor else {
+            return
+        }
+        
+        let targetTimestampNS = UInt64(targetTimestampWalkedBack * Double(NSEC_PER_SEC))
+        
+        deviceAnchorsQueue.append(targetTimestampNS)
+        if deviceAnchorsQueue.count > 1000 {
+            let val = deviceAnchorsQueue.removeFirst()
+            deviceAnchorsDictionary.removeValue(forKey: val)
+        }
+        deviceAnchorsDictionary[targetTimestampNS] = deviceAnchor.originFromAnchorTransform
+        let orientation = simd_quaternion(deviceAnchor.originFromAnchorTransform)
+        let position = deviceAnchor.originFromAnchorTransform.columns.3
+        var trackingMotion = AlvrDeviceMotion(device_id: WorldTracker.deviceIdHead, orientation: AlvrQuat(x: orientation.vector.x, y: orientation.vector.y, z: orientation.vector.z, w: orientation.vector.w), position: (position.x, position.y, position.z), linear_velocity: (0, 0, 0), angular_velocity: (0, 0, 0))
+        let targetTimestampReqestedNS = UInt64(targetTimestamp * Double(NSEC_PER_SEC))
+        let currentTimeNs = UInt64(CACurrentMediaTime() * Double(NSEC_PER_SEC))
+        //print("asking for:", targetTimestampNS, "diff:", targetTimestampReqestedNS&-targetTimestampNS, "diff2:", targetTimestampNS&-lastRequestedTimestamp, "diff3:", targetTimestampNS&-currentTimeNs)
+        EventHandler.shared.lastRequestedTimestamp = targetTimestampNS
+        alvr_send_tracking(targetTimestampNS, &trackingMotion, 1)
+    }
+    
+    func lookupDeviceAnchorFor(timestamp: UInt64) -> simd_float4x4? {
+        return deviceAnchorsDictionary[timestamp]
+    }
+    
+    func reset() {
+        
+        objc_sync_enter(deviceAnchorsLock)
+        deviceAnchorsQueue.removeAll()
+        deviceAnchorsDictionary.removeAll()
+        objc_sync_exit(deviceAnchorsLock)
+        
+    }
+}
